working on updateRewards

+++ gridworld/src/main/java/org/hanns/environments/discrete/world/objects/Tale.java
@@ -15,6 +15,6 @@ public interface Tale{
 	
 	public String getLabel();
 	
-	public int getReward();
+	public float getReward();
 
 }
